+++ rules/enforce-effect-naming-convention/examples/correct-examples-issue-24.js
@@ -1,6 +1,6 @@
 import { combine } from "effector";
 
-// Examples were found in production code-base with false-poitive detection on 0.1.2
+// Examples were found in production code-base with false-positive detection on 0.1.2
 // https://github.com/igorkamyshev/eslint-plugin-effector/issues/24
 
 const $sourceOrDefault = combine($source, $allSources, (source, allSources) => {