+++ rules/enforce-store-naming-convention/prefix/examples/correct-examples-issue-23.js
@@ -1,4 +1,4 @@
-// Examples were found in production code-base with false-poitive detection on 0.1.2
+// Examples were found in production code-base with false-positive detection on 0.1.2
 // https://github.com/igorkamyshev/eslint-plugin-effector/issues/23
 
 forward({