+++ rules/no-useless-methods/examples/correct-examples-issue-74.js
@@ -6,7 +6,7 @@ const $nameValue = createStore("");
 
 const nameInputChanged = createEvent();
 
-// Examples were found in production code-base with false-poitive detection on 0.4.1
+// Examples were found in production code-base with false-positive detection on 0.4.1
 // https://github.com/igorkamyshev/eslint-plugin-effector/issues/74
 
 $nameValue.on(