+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 
 - Add new rule `keep-options-order` ([PR #81](https://github.com/effector/eslint-plugin/pull/81))
 - Add new rule `no-forward` ([PR #82](https://github.com/effector/eslint-plugin/pull/82))
-- Fix false-positive on nested obejct with store in `no-duplicate-on` ([PR #86](https://github.com/effector/eslint-plugin/pull/86))
+- Fix false-positive on nested object with store in `no-duplicate-on` ([PR #86](https://github.com/effector/eslint-plugin/pull/86))
 - Fix possible exception on WIP code in some IDEs in `no-duplicate-on`
 
 ## v0.5.2